Cleanup test sfpu binary and remove where standalone test completely -> unify in ternary test
