+++ tt_metal/tt-llk/tests/python_tests/helpers/llk_params.py
@@ -285,7 +285,7 @@ class MathOperation(Enum):
     # =============================================================================
     # SFPU TERNARY OPERATIONS
     # =============================================================================
-    SfpuWhere = OpSpec("WHERE", MathOpType.SFPU_TERNARY)
+    SfpuWhere = OpSpec("where", MathOpType.SFPU_TERNARY)
     TTNNWhere = SfpuWhere
     SfpuAddcmul = OpSpec("addcmul", MathOpType.SFPU_TERNARY)
     SfpuAddcdiv = OpSpec("addcdiv", MathOpType.SFPU_TERNARY)
