+++ tt_metal/tt-llk/tests/python_tests/test_sfpu_binary.py
@@ -40,6 +40,10 @@ from helpers.test_variant_parameters import (
 from helpers.tilize_untilize import tilize
 from helpers.utils import passed_test
 
+# =============================================================================
+# Shared skip helpers
+# =============================================================================
+
 
 def _skip_fp32_no_dest_acc(formats, dest_acc):
     """32-bit (Float32) inputs need a 32-bit dest, i.e. dest_acc=Yes."""
@@ -57,6 +61,228 @@ def _skip_bh_float16_no_dest_acc(formats, dest_acc):
         pytest.skip(
             "Float16_a isn't supported for SFPU on Blackhole without being converted to 32-bit intermediate format in dest register"
         )
+
+
+# =============================================================================
+# Shared crafted-stimuli helpers
+#
+# Several predicate/paired ops (mask, isclose, eq/ne, logsigmoid) need operand
+# tiles filled from *different* per-position data, which the default random
+# sweep can't express. These builders produce those StimuliSpecs.
+# =============================================================================
+
+# Number of faces per tile for the [64, 32] two-tile binary harness layout
+# (a 32x32 tile is 4 faces of 16x16, and input_dimensions=[64, 32] is 8 faces).
+_FACES_PER_TILE = 4
+
+
+def _paired_two_tile_spec(a_face, b_face):
+    """Fill operand tile0 (in0) and tile1 (in1) from *different* per-position data.
+
+    The harness applies one per-face distribution to every face, so a plain callable makes
+    in0 == in1. Keep `a_face` for the tile0 faces (0..3) and override the tile1 faces (4..7)
+    with `b_face`, so position p pairs as (a_face[p], b_face[p]); tilize preserves it.
+    """
+    return StimuliSpec(
+        distribution=a_face,
+        seed=0,
+        face_specs=[None] * _FACES_PER_TILE
+        + [StimuliSpec(distribution=b_face, seed=0)] * _FACES_PER_TILE,
+    )
+
+
+def _mask_stimuli_spec():
+    # mask zeroes data (in0) where mask (in1) is 0. Data and mask are separate tiles: keep
+    # data strictly non-zero (1..8) and zero ~1/3 of the mask, so a passthrough kernel fails.
+    def data_face(size, dtype, generator):
+        j = torch.arange(size, dtype=torch.float32)
+        return (1.0 + (j % 8)).to(dtype)  # 1..8, always non-zero
+
+    def mask_face(size, dtype, generator):
+        j = torch.arange(size, dtype=torch.float32)
+        return torch.where(j % 3 == 0, 0.0, 1.0).to(dtype)  # ~1/3 exact zeros
+
+    return _paired_two_tile_spec(data_face, mask_face)
+
+
+def _isclose_stimuli_spec():
+    # isclose is a predicate on paired operands (a = tile0, b = tile1). Fill the two tiles
+    # from different data so even p -> identical (isclose 1), odd p -> differ by 2.0
+    # (isclose 0); the 2.0 gap dwarfs the tolerance so the decision is unambiguous.
+    def a_face(size, dtype, generator):
+        j = torch.arange(size, dtype=torch.float32)
+        return (1.0 + (j % 8)).to(dtype)  # 1..8, strictly positive
+
+    def b_face(size, dtype, generator):
+        j = torch.arange(size, dtype=torch.float32)
+        base = 1.0 + (j % 8)
+        return (base + torch.where(j % 2 == 0, 0.0, 2.0)).to(dtype)
+
+    return _paired_two_tile_spec(a_face, b_face)
+
+
+def _eq_ne_stimuli_spec():
+    # Eq/Ne compare paired operands (a = tile0, b = tile1). Fill the two tiles so even p ->
+    # identical (Eq 1), odd p -> differ by 1.0 (Eq 0), a clean ~50/50 mix.
+    def a_face(size, dtype, generator):
+        j = torch.arange(size, dtype=torch.float32)
+        return (1.0 + (j % 8)).to(dtype)  # 1..8
+
+    def b_face(size, dtype, generator):
+        j = torch.arange(size, dtype=torch.float32)
+        base = 1.0 + (j % 8)
+        return (base + torch.where(j % 2 == 0, 0.0, 1.0)).to(dtype)
+
+    return _paired_two_tile_spec(a_face, b_face)
+
+
+def _logsigmoid_stimuli_spec():
+    # logsigmoid(x) = -softplus(-x). in1 (exp(-x)) is only read in the x > 4 branch, so
+    # restrict x to [-8, 3.9] (never uses in1) and sweep the passthrough (x < -4) and
+    # polynomial (-4 < x < 4) branches. The distribution is invoked per 16x16 face (size 256).
+    def dist(size, dtype, generator):
+        return torch.linspace(-8.0, 3.9, size).to(dtype)
+
+    return StimuliSpec(distribution=dist, seed=0)
+
+
+# =============================================================================
+# Shared driver
+#
+# Every test below (except add_top_row and the separate broadcast kernel) runs
+# through sfpu_binary(): it builds the [64, 32] two-tile stimuli, computes the
+# golden per tile-pair, and drives sources/sfpu_binary_test.cpp.
+# =============================================================================
+
+
+def sfpu_binary(
+    formats,
+    dest_acc,
+    mathop,
+    broadcast_type=None,
+    src_A_override=None,
+    spec_A=None,
+    twos_complement=False,
+):
+
+    # FP32 destination tiles occupy twice the register space. Keep four full destination
+    # blocks for those formats and four blocks of eight tiles for the remaining formats.
+    input_dimensions = [128, 128] if formats.input_format.is_32_bit() else [256, 128]
+
+    src_A, tile_cnt_A, src_B, tile_cnt_B = generate_stimuli(
+        stimuli_format_A=formats.input_format,
+        input_dimensions_A=input_dimensions,
+        stimuli_format_B=formats.input_format,
+        input_dimensions_B=input_dimensions,
+        spec_A=spec_A,
+    )
+
+    # The kernel only consumes buffer_A (operand 0 = tile 0, operand 1 = tile 1), so an
+    # explicit src_A fully controls inputs for edge cases; src_B stays random but unused.
+    if src_A_override is not None:
+        override = src_A_override.to(src_A.dtype).flatten()
+        if src_A.numel() % override.numel() != 0:
+            raise ValueError(
+                "SFPU binary override must contain a whole number of tile pairs"
+            )
+        src_A = override.repeat(src_A.numel() // override.numel())
+
+    golden_src = src_A
+    if broadcast_type is not None and broadcast_type != LlkBroadcastType.None_:
+        generate_broadcast_golden = get_golden_generator(BroadcastGolden)
+        golden_src = generate_broadcast_golden(
+            broadcast_type,
+            src_A,
+            (
+                formats.input_format
+                if formats.input_format != DataFormat.Bfp8_b
+                else DataFormat.Float16_b
+            ),
+            tile_cnt=tile_cnt_A,
+        )
+
+    generate_golden = get_golden_generator(BinarySFPUGolden)
+    golden_format = (
+        DataFormat.Float16_b
+        if formats.input_format == DataFormat.Bfp8_b
+        else formats.input_format
+    )
+    elements_per_pair = 2 * 32 * 32
+    golden_tensor = torch.cat(
+        [
+            generate_golden(
+                mathop,
+                golden_src[offset : offset + elements_per_pair],
+                0,
+                1,
+                0,
+                32,
+                [64, 32],
+                golden_format,
+            ).flatten()
+            for offset in range(0, golden_src.numel(), elements_per_pair)
+        ]
+    )
+
+    # ONLY Blackhole needs this for some reason
+    if (
+        formats.input_format in [DataFormat.Float16, DataFormat.Float32]
+        and TestConfig.CHIP_ARCH == ChipArchitecture.BLACKHOLE
+    ):
+        dest_acc = DestAccumulation.Yes
+
+    bcast = broadcast_type if broadcast_type else LlkBroadcastType.None_
+
+    num_blocks, num_tiles_in_block = get_num_blocks_and_num_tiles_in_block(
+        DestSync.Half, dest_acc, formats, input_dimensions, TILE_DIMENSIONS
+    )
+
+    configuration = TestConfig(
+        "sources/sfpu_binary_test.cpp",
+        formats,
+        templates=[
+            generate_input_dim(input_dimensions, input_dimensions),
+            MATH_OP(mathop=mathop),
+            APPROX_MODE(),
+            BROADCAST_TYPE(bcast),
+        ],
+        runtimes=[
+            TILE_COUNT(tile_cnt_A),
+            NUM_BLOCKS(num_blocks),
+            NUM_TILES_IN_BLOCK(num_tiles_in_block),
+        ],
+        variant_stimuli=StimuliConfig(
+            src_A,
+            formats.input_format,
+            src_B,
+            formats.input_format,
+            formats.output_format,
+            tile_count_A=tile_cnt_A,
+            tile_count_B=tile_cnt_B,
+            tile_count_res=tile_cnt_A,
+            twos_complement=twos_complement,
+        ),
+        dest_acc=dest_acc,
+        unpack_to_dest=formats.input_format.is_32_bit(),
+        compile_time_formats=True,
+    )
+    res_from_L1 = configuration.run().result
+
+    torch_format = format_dict[formats.output_format]
+    res_tensor = torch.tensor(res_from_L1, dtype=torch_format).flatten()
+
+    assert len(res_tensor) == len(
+        golden_tensor
+    ), "Result tensor and golden tensor are not of the same length"
+
+    assert passed_test(
+        golden_tensor, res_tensor, formats.output_format
+    ), "Assert against golden failed"
+
+
+# =============================================================================
+# Float ops
+# =============================================================================
 
 
 @parametrize(
@@ -155,6 +381,142 @@ def test_sfpu_binary_div(formats, dest_acc):
 @parametrize(
     formats=input_output_formats(
         [
+            DataFormat.Float32,
+            DataFormat.Float16,
+            DataFormat.Float16_b,
+            DataFormat.Bfp8_b,
+        ]
+    ),
+    mathop=[
+        MathOperation.SfpuBinaryMax,
+        MathOperation.SfpuBinaryMin,
+        MathOperation.SfpuBinaryFmod,
+        MathOperation.SfpuBinaryRemainder,
+    ],
+    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
+)
+def test_sfpu_binary_float_extended(formats, dest_acc, mathop):
+    # max/min (SFPSWAP) and fmod/remainder (fp32 reciprocal) binary kernels with no
+    # dedicated production BinaryOp; driven through the same in-DST harness as add/sub.
+    _skip_fp32_no_dest_acc(formats, dest_acc)
+    _skip_bh_float16_no_dest_acc(formats, dest_acc)
+
+    # fmod/remainder divide by b via a reciprocal; Bfp8_b's coarse quantization blows up
+    # the quotient for small divisors (mirrors the pow/xlogy Bfp8_b skip above).
+    if formats.input_format == DataFormat.Bfp8_b and mathop in (
+        MathOperation.SfpuBinaryFmod,
+        MathOperation.SfpuBinaryRemainder,
+    ):
+        pytest.skip("Bfp8_b is not supported for fmod/remainder coverage")
+
+    sfpu_binary(
+        formats,
+        dest_acc,
+        mathop,
+        broadcast_type=LlkBroadcastType.None_,
+    )
+
+
+@parametrize(
+    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
+    mathop=[MathOperation.SfpuMask],
+    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
+)
+def test_sfpu_binary_mask(formats, dest_acc, mathop):
+    # float mask: data at tile0, mask at tile1. Output is data where mask != 0, else 0.
+    # Crafted stimuli so the mask carries real zeros.
+    _skip_fp32_no_dest_acc(formats, dest_acc)
+
+    sfpu_binary(
+        formats,
+        dest_acc,
+        mathop,
+        broadcast_type=LlkBroadcastType.None_,
+        spec_A=_mask_stimuli_spec(),
+    )
+
+
+@parametrize(
+    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
+    mathop=[MathOperation.SfpuAtan2],
+    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
+)
+def test_sfpu_binary_atan2(formats, dest_acc, mathop):
+    # atan2(y, x): y = tile0, x = tile1. Signed [-5, 5] gives mixed signs so all quadrants
+    # (and the |y|>=|x| / x<0 branches) are exercised; minimax approximation matched under PCC.
+    _skip_fp32_no_dest_acc(formats, dest_acc)
+
+    sfpu_binary(
+        formats,
+        dest_acc,
+        mathop,
+        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0),
+    )
+
+
+@parametrize(
+    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
+    mathop=[MathOperation.SfpuElwEq, MathOperation.SfpuElwNe],
+    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
+)
+def test_sfpu_binary_eq_ne(formats, dest_acc, mathop):
+    # Eq/Ne(a, b) with a = tile0, b = tile1. Crafted paired stimuli give a non-constant 0/1
+    # golden so the equal branch is exercised (the default random sweep never is).
+    _skip_fp32_no_dest_acc(formats, dest_acc)
+
+    sfpu_binary(
+        formats,
+        dest_acc,
+        mathop,
+        spec_A=_eq_ne_stimuli_spec(),
+    )
+
+
+@parametrize(
+    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
+    mathop=[MathOperation.SfpuIsclose],
+    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
+)
+def test_sfpu_binary_isclose(formats, dest_acc, mathop):
+    # isclose(a, b) = |a - b| <= atol + rtol*|b|, a = tile0, b = tile1. torch default
+    # tolerances (fixed in the C++ dispatch); crafted stimuli give a non-constant 0/1 mix.
+    _skip_fp32_no_dest_acc(formats, dest_acc)
+
+    sfpu_binary(
+        formats,
+        dest_acc,
+        mathop,
+        spec_A=_isclose_stimuli_spec(),
+    )
+
+
+@parametrize(
+    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
+    mathop=[MathOperation.SfpuLogsigmoid],
+    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
+)
+def test_sfpu_binary_logsigmoid(formats, dest_acc, mathop):
+    # logsigmoid(x) with x = tile0. Piecewise poly/passthrough approximation matched under
+    # PCC; x swept over [-8, 3.9]. The x > 4 (-exp(-x)) branch needs a device-computed
+    # exp(-x) operand the shared harness can't provide, left to a future driver.
+    _skip_fp32_no_dest_acc(formats, dest_acc)
+
+    sfpu_binary(
+        formats,
+        dest_acc,
+        mathop,
+        spec_A=_logsigmoid_stimuli_spec(),
+    )
+
+
+# =============================================================================
+# Integer ops
+# =============================================================================
+
+
+@parametrize(
+    formats=input_output_formats(
+        [
             DataFormat.Int32,
         ]
     ),
@@ -183,10 +545,100 @@ def test_sfpu_binary_int(
     )
 
 
-# Deterministic edge-case coverage for the integer shift ops: shift amounts outside
-# [0, 31] -> 0, arithmetic right-shift sign-extends, negatives shift correctly. INT32_MIN
-# is excluded (sign-magnitude Dst can't represent -2^31); see the xfail test below and
-# SFPU_INT32_SHIFT.md.
+@parametrize(
+    formats=input_output_formats([DataFormat.Int32]),
+    mathop=[
+        MathOperation.SfpuBitwiseAnd,
+        MathOperation.SfpuBitwiseOr,
+        MathOperation.SfpuBitwiseXor,
+    ],
+    dest_acc=[DestAccumulation.Yes],
+)
+def test_sfpu_binary_bitwise(formats, dest_acc, mathop):
+    # int32 bitwise AND/OR/XOR: exact on the full default int range.
+    sfpu_binary(formats, dest_acc, mathop)
+
+
+# Ops whose kernel interprets DST as unsigned; run them under UInt32 (the rest are Int32).
+_UINT32_BINARY_OPS = {
+    MathOperation.SfpuMaxUint32,
+    MathOperation.SfpuMinUint32,
+    MathOperation.SfpuRemainderUint32,
+}
+
+# int/uint binary ops sharing the same driver: dest_acc=Yes, single-format, and a per-op
+# uniform positive stimuli range. Ranges keep operands (and results) non-negative and small
+# enough to round-trip the sign-magnitude Dst packer plus any int->fp32 reciprocal the
+# kernel uses. mathop -> (low, high).
+_INT_BINARY_STIMULI = {
+    # trunc/floor division < 2**24: exact int->fp32 reciprocal, trunc == floor, and the
+    # sign-magnitude pack path can't round-trip the negatives these kernels would emit.
+    MathOperation.SfpuDivInt32: (1.0, 8_000_000.0),
+    MathOperation.SfpuDivInt32Floor: (1.0, 8_000_000.0),
+    # binary-GCD on raw int32 bits (exact): strictly positive within the 31-bit budget.
+    MathOperation.SfpuGcd: (1.0, 100_000.0),
+    # lcm abs()es both operands and assumes |a|, |b| < 2**15.
+    MathOperation.SfpuLcm: (1.0, 20_000.0),
+    # int32 multiply low-32: operands < ~46340 so the product stays < 2**31 (non-negative).
+    MathOperation.SfpuMulInt32: (1.0, 40_000.0),
+    # int32/uint32 max/min via SFPSWAP: non-negative so signed/unsigned agree and round-trip.
+    MathOperation.SfpuMaxInt32: (0.0, 1_000_000.0),
+    MathOperation.SfpuMinInt32: (0.0, 1_000_000.0),
+    MathOperation.SfpuMaxUint32: (0.0, 1_000_000.0),
+    MathOperation.SfpuMinUint32: (0.0, 1_000_000.0),
+    # remainder/fmod: non-negative operands, divisor >= 1 so every convention agrees;
+    # kept < 2**24 for the exact int->fp32 reciprocal the quotient uses.
+    MathOperation.SfpuRemainderInt32: (1.0, 10_000.0),
+    MathOperation.SfpuFmodInt32: (1.0, 10_000.0),
+    MathOperation.SfpuRemainderUint32: (1.0, 10_000.0),
+}
+
+
+@parametrize(
+    mathop=list(_INT_BINARY_STIMULI),
+    dest_acc=[DestAccumulation.Yes],
+)
+def test_sfpu_binary_int_uniform(mathop, dest_acc):
+    int_format = DataFormat.UInt32 if mathop in _UINT32_BINARY_OPS else DataFormat.Int32
+    formats = InputOutputFormat(int_format, int_format)
+    low, high = _INT_BINARY_STIMULI[mathop]
+    sfpu_binary(
+        formats,
+        dest_acc,
+        mathop,
+        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=low, high=high),
+    )
+
+
+@parametrize(
+    formats=input_output_formats([DataFormat.Int32]),
+    mathop=[MathOperation.SfpuRsubInt32],
+    dest_acc=[DestAccumulation.Yes],
+)
+def test_sfpu_binary_rsub_int32(formats, dest_acc, mathop):
+    sfpu_binary(formats, dest_acc, mathop, twos_complement=True)
+
+
+@parametrize(
+    formats=input_output_formats([DataFormat.Int32]),
+    mathop=[MathOperation.SfpuEqInt, MathOperation.SfpuNeInt],
+    dest_acc=[DestAccumulation.Yes],
+)
+def test_sfpu_binary_eq_ne_int(formats, dest_acc, mathop):
+    # int32 eq/ne via calculate_binary_eq_int (exact 0/1 over the raw INT32 dest bits).
+    # Reuse the paired eq/ne stimuli so ~50% of positions compare equal — the equal branch
+    # a plain random int sweep would essentially never hit.
+    sfpu_binary(formats, dest_acc, mathop, spec_A=_eq_ne_stimuli_spec())
+
+
+# =============================================================================
+# Integer shift edge cases
+#
+# Deterministic edge-case coverage for the integer shift ops: shift amounts
+# outside [0, 31] -> 0, arithmetic right-shift sign-extends, negatives shift
+# correctly. INT32_MIN is excluded (sign-magnitude Dst can't represent -2^31);
+# see the xfail test below and SFPU_INT32_SHIFT.md.
+# =============================================================================
 
 _INT32_MIN = -(2**31)
 
@@ -342,296 +794,13 @@ def test_sfpu_binary_int_shift_int32_min_unsupported(
     )
 
 
-@parametrize(
-    formats=input_output_formats(
-        [
-            DataFormat.Float32,
-            DataFormat.Float16,
-            DataFormat.Float16_b,
-            DataFormat.Bfp8_b,
-        ]
-    ),
-    mathop=[
-        MathOperation.SfpuBinaryMax,
-        MathOperation.SfpuBinaryMin,
-        MathOperation.SfpuBinaryFmod,
-        MathOperation.SfpuBinaryRemainder,
-    ],
-    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
-)
-def test_sfpu_binary_float_extended(formats, dest_acc, mathop):
-    # max/min (SFPSWAP) and fmod/remainder (fp32 reciprocal) binary kernels with no
-    # dedicated production BinaryOp; driven through the same in-DST harness as add/sub.
-    _skip_fp32_no_dest_acc(formats, dest_acc)
-    _skip_bh_float16_no_dest_acc(formats, dest_acc)
-
-    # fmod/remainder divide by b via a reciprocal; Bfp8_b's coarse quantization blows up
-    # the quotient for small divisors (mirrors the pow/xlogy Bfp8_b skip above).
-    if formats.input_format == DataFormat.Bfp8_b and mathop in (
-        MathOperation.SfpuBinaryFmod,
-        MathOperation.SfpuBinaryRemainder,
-    ):
-        pytest.skip("Bfp8_b is not supported for fmod/remainder coverage")
-
-    sfpu_binary(
-        formats,
-        dest_acc,
-        mathop,
-        broadcast_type=LlkBroadcastType.None_,
-    )
-
-
-@parametrize(
-    formats=input_output_formats([DataFormat.Int32]),
-    mathop=[
-        MathOperation.SfpuBitwiseAnd,
-        MathOperation.SfpuBitwiseOr,
-        MathOperation.SfpuBitwiseXor,
-    ],
-    dest_acc=[DestAccumulation.Yes],
-)
-def test_sfpu_binary_bitwise(formats, dest_acc, mathop):
-    # int32 bitwise AND/OR/XOR: exact on the full default int range.
-    sfpu_binary(formats, dest_acc, mathop)
-
-
-# Ops whose kernel interprets DST as unsigned; run them under UInt32 (the rest are Int32).
-_UINT32_BINARY_OPS = {
-    MathOperation.SfpuMaxUint32,
-    MathOperation.SfpuMinUint32,
-    MathOperation.SfpuRemainderUint32,
-}
-
-# int/uint binary ops sharing the same driver: dest_acc=Yes, single-format, and a per-op
-# uniform positive stimuli range. Ranges keep operands (and results) non-negative and small
-# enough to round-trip the sign-magnitude Dst packer plus any int->fp32 reciprocal the
-# kernel uses. mathop -> (low, high).
-_INT_BINARY_STIMULI = {
-    # trunc/floor division < 2**24: exact int->fp32 reciprocal, trunc == floor, and the
-    # sign-magnitude pack path can't round-trip the negatives these kernels would emit.
-    MathOperation.SfpuDivInt32: (1.0, 8_000_000.0),
-    MathOperation.SfpuDivInt32Floor: (1.0, 8_000_000.0),
-    # binary-GCD on raw int32 bits (exact): strictly positive within the 31-bit budget.
-    MathOperation.SfpuGcd: (1.0, 100_000.0),
-    # lcm abs()es both operands and assumes |a|, |b| < 2**15.
-    MathOperation.SfpuLcm: (1.0, 20_000.0),
-    # int32 multiply low-32: operands < ~46340 so the product stays < 2**31 (non-negative).
-    MathOperation.SfpuMulInt32: (1.0, 40_000.0),
-    # int32/uint32 max/min via SFPSWAP: non-negative so signed/unsigned agree and round-trip.
-    MathOperation.SfpuMaxInt32: (0.0, 1_000_000.0),
-    MathOperation.SfpuMinInt32: (0.0, 1_000_000.0),
-    MathOperation.SfpuMaxUint32: (0.0, 1_000_000.0),
-    MathOperation.SfpuMinUint32: (0.0, 1_000_000.0),
-    # remainder/fmod: non-negative operands, divisor >= 1 so every convention agrees;
-    # kept < 2**24 for the exact int->fp32 reciprocal the quotient uses.
-    MathOperation.SfpuRemainderInt32: (1.0, 10_000.0),
-    MathOperation.SfpuFmodInt32: (1.0, 10_000.0),
-    MathOperation.SfpuRemainderUint32: (1.0, 10_000.0),
-}
-
-
-@parametrize(
-    mathop=list(_INT_BINARY_STIMULI),
-    dest_acc=[DestAccumulation.Yes],
-)
-def test_sfpu_binary_int_uniform(mathop, dest_acc):
-    int_format = DataFormat.UInt32 if mathop in _UINT32_BINARY_OPS else DataFormat.Int32
-    formats = InputOutputFormat(int_format, int_format)
-    low, high = _INT_BINARY_STIMULI[mathop]
-    sfpu_binary(
-        formats,
-        dest_acc,
-        mathop,
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=low, high=high),
-    )
-
-
-@parametrize(
-    formats=input_output_formats([DataFormat.Int32]),
-    mathop=[MathOperation.SfpuRsubInt32],
-    dest_acc=[DestAccumulation.Yes],
-)
-def test_sfpu_binary_rsub_int32(formats, dest_acc, mathop):
-    sfpu_binary(formats, dest_acc, mathop, twos_complement=True)
-
-
-# Number of faces per tile for the [64, 32] two-tile binary harness layout
-# (a 32x32 tile is 4 faces of 16x16, and input_dimensions=[64, 32] is 8 faces).
-_FACES_PER_TILE = 4
-
-
-def _paired_two_tile_spec(a_face, b_face):
-    """Fill operand tile0 (in0) and tile1 (in1) from *different* per-position data.
-
-    The harness applies one per-face distribution to every face, so a plain callable makes
-    in0 == in1. Keep `a_face` for the tile0 faces (0..3) and override the tile1 faces (4..7)
-    with `b_face`, so position p pairs as (a_face[p], b_face[p]); tilize preserves it.
-    """
-    return StimuliSpec(
-        distribution=a_face,
-        seed=0,
-        face_specs=[None] * _FACES_PER_TILE
-        + [StimuliSpec(distribution=b_face, seed=0)] * _FACES_PER_TILE,
-    )
-
-
-def _mask_stimuli_spec():
-    # mask zeroes data (in0) where mask (in1) is 0. Data and mask are separate tiles: keep
-    # data strictly non-zero (1..8) and zero ~1/3 of the mask, so a passthrough kernel fails.
-    def data_face(size, dtype, generator):
-        j = torch.arange(size, dtype=torch.float32)
-        return (1.0 + (j % 8)).to(dtype)  # 1..8, always non-zero
-
-    def mask_face(size, dtype, generator):
-        j = torch.arange(size, dtype=torch.float32)
-        return torch.where(j % 3 == 0, 0.0, 1.0).to(dtype)  # ~1/3 exact zeros
-
-    return _paired_two_tile_spec(data_face, mask_face)
-
-
-@parametrize(
-    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
-    mathop=[MathOperation.SfpuMask],
-    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
-)
-def test_sfpu_binary_mask(formats, dest_acc, mathop):
-    # float mask: data at tile0, mask at tile1. Output is data where mask != 0, else 0.
-    # Crafted stimuli so the mask carries real zeros.
-    _skip_fp32_no_dest_acc(formats, dest_acc)
-
-    sfpu_binary(
-        formats,
-        dest_acc,
-        mathop,
-        broadcast_type=LlkBroadcastType.None_,
-        spec_A=_mask_stimuli_spec(),
-    )
-
-
-@parametrize(
-    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
-    mathop=[MathOperation.SfpuAtan2],
-    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
-)
-def test_sfpu_binary_atan2(formats, dest_acc, mathop):
-    # atan2(y, x): y = tile0, x = tile1. Signed [-5, 5] gives mixed signs so all quadrants
-    # (and the |y|>=|x| / x<0 branches) are exercised; minimax approximation matched under PCC.
-    _skip_fp32_no_dest_acc(formats, dest_acc)
-
-    sfpu_binary(
-        formats,
-        dest_acc,
-        mathop,
-        spec_A=StimuliSpec(distribution=DistributionKind.UNIFORM, low=-5.0, high=5.0),
-    )
-
-
-@parametrize(
-    formats=input_output_formats([DataFormat.Int32]),
-    mathop=[MathOperation.SfpuEqInt, MathOperation.SfpuNeInt],
-    dest_acc=[DestAccumulation.Yes],
-)
-def test_sfpu_binary_eq_ne_int(formats, dest_acc, mathop):
-    # int32 eq/ne via calculate_binary_eq_int (exact 0/1 over the raw INT32 dest bits).
-    # Reuse the paired eq/ne stimuli so ~50% of positions compare equal — the equal branch
-    # a plain random int sweep would essentially never hit.
-    sfpu_binary(formats, dest_acc, mathop, spec_A=_eq_ne_stimuli_spec())
-
-
-def _isclose_stimuli_spec():
-    # isclose is a predicate on paired operands (a = tile0, b = tile1). Fill the two tiles
-    # from different data so even p -> identical (isclose 1), odd p -> differ by 2.0
-    # (isclose 0); the 2.0 gap dwarfs the tolerance so the decision is unambiguous.
-    def a_face(size, dtype, generator):
-        j = torch.arange(size, dtype=torch.float32)
-        return (1.0 + (j % 8)).to(dtype)  # 1..8, strictly positive
-
-    def b_face(size, dtype, generator):
-        j = torch.arange(size, dtype=torch.float32)
-        base = 1.0 + (j % 8)
-        return (base + torch.where(j % 2 == 0, 0.0, 2.0)).to(dtype)
-
-    return _paired_two_tile_spec(a_face, b_face)
-
-
-@parametrize(
-    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
-    mathop=[MathOperation.SfpuIsclose],
-    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
-)
-def test_sfpu_binary_isclose(formats, dest_acc, mathop):
-    # isclose(a, b) = |a - b| <= atol + rtol*|b|, a = tile0, b = tile1. torch default
-    # tolerances (fixed in the C++ dispatch); crafted stimuli give a non-constant 0/1 mix.
-    _skip_fp32_no_dest_acc(formats, dest_acc)
-
-    sfpu_binary(
-        formats,
-        dest_acc,
-        mathop,
-        spec_A=_isclose_stimuli_spec(),
-    )
-
-
-def _eq_ne_stimuli_spec():
-    # Eq/Ne compare paired operands (a = tile0, b = tile1). Fill the two tiles so even p ->
-    # identical (Eq 1), odd p -> differ by 1.0 (Eq 0), a clean ~50/50 mix.
-    def a_face(size, dtype, generator):
-        j = torch.arange(size, dtype=torch.float32)
-        return (1.0 + (j % 8)).to(dtype)  # 1..8
-
-    def b_face(size, dtype, generator):
-        j = torch.arange(size, dtype=torch.float32)
-        base = 1.0 + (j % 8)
-        return (base + torch.where(j % 2 == 0, 0.0, 1.0)).to(dtype)
-
-    return _paired_two_tile_spec(a_face, b_face)
-
-
-@parametrize(
-    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
-    mathop=[MathOperation.SfpuElwEq, MathOperation.SfpuElwNe],
-    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
-)
-def test_sfpu_binary_eq_ne(formats, dest_acc, mathop):
-    # Eq/Ne(a, b) with a = tile0, b = tile1. Crafted paired stimuli give a non-constant 0/1
-    # golden so the equal branch is exercised (the default random sweep never is).
-    _skip_fp32_no_dest_acc(formats, dest_acc)
-
-    sfpu_binary(
-        formats,
-        dest_acc,
-        mathop,
-        spec_A=_eq_ne_stimuli_spec(),
-    )
-
-
-def _logsigmoid_stimuli_spec():
-    # logsigmoid(x) = -softplus(-x). in1 (exp(-x)) is only read in the x > 4 branch, so
-    # restrict x to [-8, 3.9] (never uses in1) and sweep the passthrough (x < -4) and
-    # polynomial (-4 < x < 4) branches. The distribution is invoked per 16x16 face (size 256).
-    def dist(size, dtype, generator):
-        return torch.linspace(-8.0, 3.9, size).to(dtype)
-
-    return StimuliSpec(distribution=dist, seed=0)
-
-
-@parametrize(
-    formats=input_output_formats([DataFormat.Float16_b, DataFormat.Float32]),
-    mathop=[MathOperation.SfpuLogsigmoid],
-    dest_acc=[DestAccumulation.No, DestAccumulation.Yes],
-)
-def test_sfpu_binary_logsigmoid(formats, dest_acc, mathop):
-    # logsigmoid(x) with x = tile0. Piecewise poly/passthrough approximation matched under
-    # PCC; x swept over [-8, 3.9]. The x > 4 (-exp(-x)) branch needs a device-computed
-    # exp(-x) operand the shared harness can't provide, left to a future driver.
-    _skip_fp32_no_dest_acc(formats, dest_acc)
-
-    sfpu_binary(
-        formats,
-        dest_acc,
-        mathop,
-        spec_A=_logsigmoid_stimuli_spec(),
-    )
+# =============================================================================
+# add_top_row
+#
+# Own inline driver: uses sources/sfpu_binary_test.cpp with a single [64, 32]
+# tile pair, disabled format inference, and a golden reshaped to the top-row
+# layout — different enough from sfpu_binary() to keep separate.
+# =============================================================================
 
 
 @parametrize(
@@ -725,136 +894,14 @@ def test_sfpu_binary_add_top_row(formats, dest_acc, mathop):
     ), "Assert against golden failed"
 
 
-def sfpu_binary(
-    formats,
-    dest_acc,
-    mathop,
-    broadcast_type=None,
-    src_A_override=None,
-    spec_A=None,
-    spec_B=None,
-    twos_complement=False,
-):
-
-    # FP32 destination tiles occupy twice the register space. Keep four full destination
-    # blocks for those formats and four blocks of eight tiles for the remaining formats.
-    input_dimensions = [128, 128] if formats.input_format.is_32_bit() else [256, 128]
-
-    src_A, tile_cnt_A, src_B, tile_cnt_B = generate_stimuli(
-        stimuli_format_A=formats.input_format,
-        input_dimensions_A=input_dimensions,
-        stimuli_format_B=formats.input_format,
-        input_dimensions_B=input_dimensions,
-        spec_A=spec_A,
-        spec_B=spec_B,
-    )
-
-    # The kernel only consumes buffer_A (operand 0 = tile 0, operand 1 = tile 1), so an
-    # explicit src_A fully controls inputs for edge cases; src_B stays random but unused.
-    if src_A_override is not None:
-        override = src_A_override.to(src_A.dtype).flatten()
-        if src_A.numel() % override.numel() != 0:
-            raise ValueError(
-                "SFPU binary override must contain a whole number of tile pairs"
-            )
-        src_A = override.repeat(src_A.numel() // override.numel())
-
-    golden_src = src_A
-    if broadcast_type is not None and broadcast_type != LlkBroadcastType.None_:
-        generate_broadcast_golden = get_golden_generator(BroadcastGolden)
-        golden_src = generate_broadcast_golden(
-            broadcast_type,
-            src_A,
-            (
-                formats.input_format
-                if formats.input_format != DataFormat.Bfp8_b
-                else DataFormat.Float16_b
-            ),
-            tile_cnt=tile_cnt_A,
-        )
-
-    generate_golden = get_golden_generator(BinarySFPUGolden)
-    golden_format = (
-        DataFormat.Float16_b
-        if formats.input_format == DataFormat.Bfp8_b
-        else formats.input_format
-    )
-    elements_per_pair = 2 * 32 * 32
-    golden_tensor = torch.cat(
-        [
-            generate_golden(
-                mathop,
-                golden_src[offset : offset + elements_per_pair],
-                0,
-                1,
-                0,
-                32,
-                [64, 32],
-                golden_format,
-            ).flatten()
-            for offset in range(0, golden_src.numel(), elements_per_pair)
-        ]
-    )
-
-    # ONLY Blackhole needs this for some reason
-    if (
-        formats.input_format in [DataFormat.Float16, DataFormat.Float32]
-        and TestConfig.CHIP_ARCH == ChipArchitecture.BLACKHOLE
-    ):
-        dest_acc = DestAccumulation.Yes
-
-    bcast = broadcast_type if broadcast_type else LlkBroadcastType.None_
-
-    num_blocks, num_tiles_in_block = get_num_blocks_and_num_tiles_in_block(
-        DestSync.Half, dest_acc, formats, input_dimensions, TILE_DIMENSIONS
-    )
-
-    configuration = TestConfig(
-        "sources/sfpu_binary_test.cpp",
-        formats,
-        templates=[
-            generate_input_dim(input_dimensions, input_dimensions),
-            MATH_OP(mathop=mathop),
-            APPROX_MODE(),
-            BROADCAST_TYPE(bcast),
-        ],
-        runtimes=[
-            TILE_COUNT(tile_cnt_A),
-            NUM_BLOCKS(num_blocks),
-            NUM_TILES_IN_BLOCK(num_tiles_in_block),
-        ],
-        variant_stimuli=StimuliConfig(
-            src_A,
-            formats.input_format,
-            src_B,
-            formats.input_format,
-            formats.output_format,
-            tile_count_A=tile_cnt_A,
-            tile_count_B=tile_cnt_B,
-            tile_count_res=tile_cnt_A,
-            twos_complement=twos_complement,
-        ),
-        dest_acc=dest_acc,
-        unpack_to_dest=formats.input_format.is_32_bit(),
-        compile_time_formats=True,
-    )
-    res_from_L1 = configuration.run().result
-
-    torch_format = format_dict[formats.output_format]
-    res_tensor = torch.tensor(res_from_L1, dtype=torch_format).flatten()
-
-    assert len(res_tensor) == len(
-        golden_tensor
-    ), "Result tensor and golden tensor are not of the same length"
-
-    assert passed_test(
-        golden_tensor, res_tensor, formats.output_format
-    ), "Assert against golden failed"
-
-
-# SFPU binary with row/column broadcast (BCAST_COL / BCAST_ROW). Uses its own 3-tile
-# kernel source (sources/sfpu_binary_bcast_test.cpp) with a custom init and full-tile
-# driver; InstrModLoadStore::DEFAULT works for any float dest format (compute is FP32).
+# =============================================================================
+# Broadcast kernel
+#
+# SFPU binary with row/column broadcast (BCAST_COL / BCAST_ROW). Uses its own
+# 3-tile kernel source (sources/sfpu_binary_bcast_test.cpp) with a custom init
+# and full-tile driver; InstrModLoadStore::DEFAULT works for any float dest
+# format (compute is FP32).
+# =============================================================================
 
 
 class BroadcastType(Enum):
